- removed unused pass-through custodian and agency based routes that are not tied to any business logic - consolidated all route functions into their respective top level files for easier understanding, including ssi - scanned app for all used routes (minus webhook) and removed the cruft - general cleanup of app.js to remove no longer applicable items
