+++ routes/health.js
@@ -1,12 +1,22 @@
 const router = require('koa-router')();
-const { getHealth } = require('./ssi/sc-admin');
 const { logRoutes } = require('../src/util');
+const { Credentials, CredentialsServiceClient, ProviderServiceClient } = require('@trinsic/service-clients');
+
+const { STREETCRED_API_KEY } = require('../config');
+
+const client = new CredentialsServiceClient(new Credentials(STREETCRED_API_KEY));
 
 module.exports = ({ psql }) => {
   router.get('/health/streetcred', getHealth);
   router.get('/health', healthCheck);
   router.get('/', healthCheck);
   router.get('/version', getVersion);
+
+  async function getHealth(ctx, next) {
+    const result = await client.health();
+    ctx.body = result;
+    return ctx;
+  }
 
   async function healthCheck(ctx, next) {
     const [{ version }] = await psql.select(psql.raw('version()'));
@@ -23,7 +33,7 @@ module.exports = ({ psql }) => {
   // logRoutes(router);
 
   return router;
-}
+};
 
 async function getVersion(ctx) {
   ctx.body = {
@@ -32,7 +42,7 @@ async function getVersion(ctx) {
   };
 
   ctx.status = 200;
-  
+
   logRoutes(router);
   return router;
 }
